Update `TopeCerrado` position and refactor `Prueba_servos.java` for dynamic testing
- **ServosSubsystem_Autonomous.java**:
    - Update `TopeCerrado()` position from `0` to `0.35`.

- **Prueba_servos.java**:
    - Add static variables `Position` and `hola` for real-time dashboard tuning.
    - Remove hardcoded `set(0)` and `setInverted(true)` from `init()`.
    - Update `loop()` to dynamically set inversion based on `hola` and use the `Position` variable when button B is pressed.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/COMANDOS_AUTONOMOS/SetTurretPositionCommand.java
@@ -6,17 +6,17 @@ import Subsistemas.TurretSubsystem_Autonomous;
 
 public class SetTurretPositionCommand extends CommandBase {
     private final TurretSubsystem_Autonomous turret;
-    private final int targetPosition;
+    private final double targetPosition;
     private final boolean waitForTarget;
 
     /**
      * Comando para mover la torreta a una posición
      * @param turret Subsistema de la torreta
-     * @param targetPosition Posición objetivo en ticks (ej: 500, -1000, 0)
+     * @param targetPosition Posición objetivo en ticks (puede ser decimal)
      * @param waitForTarget Si debe esperar a llegar a la posición
      */
     public SetTurretPositionCommand(TurretSubsystem_Autonomous turret,
-                                    int targetPosition,
+                                    double targetPosition,
                                     boolean waitForTarget) {
         this.turret = turret;
         this.targetPosition = targetPosition;
@@ -24,8 +24,8 @@ public class SetTurretPositionCommand extends CommandBase {
         addRequirements(turret);
     }
 
-    // Constructor simplificado que espera por defecto
-    public SetTurretPositionCommand(TurretSubsystem_Autonomous turret, int targetPosition) {
+    // Constructor simplificado
+    public SetTurretPositionCommand(TurretSubsystem_Autonomous turret, double targetPosition) {
         this(turret, targetPosition, true);
     }
